fix(expose): enforce session preempt boundary
