+++ crates/prodex-app/src/expose.rs
@@ -74,6 +74,9 @@ mod mcp_tests;
 #[path = "expose/run_manager_tests.rs"]
 mod run_manager_tests;
 #[cfg(all(test, unix))]
+#[path = "expose/session_preempt_live_tests.rs"]
+mod session_preempt_live_tests;
+#[cfg(all(test, unix))]
 #[path = "expose/session_preempt_tests.rs"]
 mod session_preempt_tests;
 #[cfg(all(test, unix))]
